refactor: Separate card config interface to its own module
