+++ src/config/service.ts
@@ -1,4 +1,10 @@
-import { Service } from '../types'
+export interface Service {
+  domain: string
+  service: string
+  data?: {
+    [key: string]: string
+  }
+}
 
 export default function parseServie(config: false | Service): Service {
   if (!config) {
